Added abstraction for a custom RequestHeaderParserInterfaces
