+++ tests/ServerTest.php
@@ -1256,26 +1256,6 @@ class ServerTest extends TestCase
         $this->assertContains("Error 505: HTTP Version not supported", $buffer);
     }
 
-    public function testMaxRequestHeaderOptionShouldFailOnInvalidType()
-    {
-        $options = array();
-        $options['max_header_size'] = 'abc';
-
-        $this->setExpectedException('InvalidArgumentException', 'Parameter "max_header_size" expected to be an integer.');
-
-        new Server($this->expectCallableNever(), $options);
-    }
-
-    public function testMaxRequestHeaderOptionShouldFailOnNegativeValue()
-    {
-        $options = array();
-        $options['max_header_size'] = -1024;
-
-        $this->setExpectedException('InvalidArgumentException', 'Parameter "max_header_size" expected to be a positive value.');
-
-        new Server($this->expectCallableNever(), $options);
-    }
-
     public function testRequestHeaderOverflowWithDefaultValueWillEmitErrorAndSendErrorResponse()
     {
         $defaultMaxHeaderSize = 4096;
@@ -1303,79 +1283,6 @@ class ServerTest extends TestCase
 
         $data = "GET / HTTP/1.1\r\nHost: example.com\r\nConnection: close\r\nX-DATA: ";
         $data .= str_repeat('A', $defaultMaxHeaderSize + 1 - strlen($data)) . "\r\n\r\n";
-        $this->connection->emit('data', array($data));
-
-        $this->assertInstanceOf('OverflowException', $error);
-
-        $this->assertContains("HTTP/1.1 431 Request Header Fields Too Large\r\n", $buffer);
-        $this->assertContains("\r\n\r\nError 431: Request Header Fields Too Large", $buffer);
-    }
-
-    public function testRequestHeaderOverflowWithCustomValue()
-    {
-        $maxHeaderSize = 1024 * 16;
-
-        $options = array();
-        $options['max_header_size'] = $maxHeaderSize;
-
-        $server = new Server(function (ServerRequestInterface $request) {
-            return new Response(200, array());
-        }, $options);
-
-        $buffer = '';
-
-        $this->connection
-            ->expects($this->any())
-            ->method('write')
-            ->will(
-                $this->returnCallback(
-                    function ($data) use (&$buffer) {
-                        $buffer .= $data;
-                    }
-                )
-            );
-
-        $server->listen($this->socket);
-        $this->socket->emit('connection', array($this->connection));
-
-        $data = "GET / HTTP/1.1\r\nHost: example.com\r\nConnection: close\r\nX-DATA: ";
-        $data .= str_repeat('A', $maxHeaderSize - strlen($data)) . "\r\n\r\n";
-        $this->connection->emit('data', array($data));
-
-        $this->assertStringStartsWith("HTTP/1.1 200 OK\r\n", $buffer);
-    }
-
-    public function testRequestHeaderOverflowWithCustomValueWillEmitErrorAndSendErrorResponse()
-    {
-        $maxHeaderSize = 1024 * 16;
-
-        $options = array();
-        $options['max_header_size'] = $maxHeaderSize;
-
-        $error = null;
-        $server = new Server($this->expectCallableNever(), $options);
-        $server->on('error', function ($message) use (&$error) {
-            $error = $message;
-        });
-
-        $buffer = '';
-
-        $this->connection
-            ->expects($this->any())
-            ->method('write')
-            ->will(
-                $this->returnCallback(
-                    function ($data) use (&$buffer) {
-                        $buffer .= $data;
-                    }
-                )
-            );
-
-        $server->listen($this->socket);
-        $this->socket->emit('connection', array($this->connection));
-
-        $data = "GET / HTTP/1.1\r\nHost: example.com\r\nConnection: close\r\nX-DATA: ";
-        $data .= str_repeat('A', $maxHeaderSize + 1 - strlen($data)) . "\r\n\r\n";
         $this->connection->emit('data', array($data));
 
         $this->assertInstanceOf('OverflowException', $error);
